Se crea interfaz crear producto y se elimina el type que existia, asi se envia de manera organizada a la carpeta correspondiente, el formulario de la pagina crear producto tambien se envia a su carpeta correspondiente para mantener un codigo ordenado, la pagina crear producto queda mas limpia y ademas la funcion de fetch se envia a la carpeta actions, aunque por ahora no se puede usar fetcher
